requirements.txt

# Para la solución de General/Solucion4.py
PyCryptodome
# Para el ejercicio de General/Solucion6
pwntools
# Para el ejercicio de General/XOR/Solucion5
Pillow